refactor independent functions into utils
