+++ src/phillips_dessouky.rs
@@ -7,11 +7,10 @@ use pathfinding::directed::edmonds_karp::{
     edmonds_karp
 };
 
-use std::time::{
-    Instant,
-    Duration,
-};
+use std::time::Instant;
 use log::info;
+
+use crate::utils;
 
 
 #[pyclass]
@@ -46,7 +45,7 @@ impl PhillipsDessouky {
             ((*from, *to), OrderedFloat(*cap))
         }).collect();
         let profiling_end = Instant::now();
-        info!("PROFILING Rust_PhillipsDessouky::max_flow scaling to OrderedFloat time: {:.10}s", PhillipsDessouky::profile_duration(profiling_start, profiling_end));
+        info!("PROFILING Rust_PhillipsDessouky::max_flow scaling to OrderedFloat time: {:.10}s", utils::profile_duration(profiling_start, profiling_end));
 
         let profiling_start = Instant::now();
         let (flows, _max_flow, _min_cut) = edmonds_karp::<_, _, _, SparseCapacity<_>>(
@@ -56,29 +55,15 @@ impl PhillipsDessouky {
             edges_edmonds_karp,
         );
         let profiling_end = Instant::now();
-        info!("PROFILING Rust_PhillipsDessouky::max_flow edmonds_karp time: {:.10}s", PhillipsDessouky::profile_duration(profiling_start, profiling_end));
+        info!("PROFILING Rust_PhillipsDessouky::max_flow edmonds_karp time: {:.10}s", utils::profile_duration(profiling_start, profiling_end));
 
         let profiling_start = Instant::now();
         let flows_f64: Vec<((u32, u32), f64)> = flows.iter().map(|((from, to), flow)| {
             ((*from, *to), flow.into_inner())
         }).collect();
         let profiling_end = Instant::now();
-        info!("PROFILING Rust_PhillipsDessouky::max_flow reformat OrderedFloat to f64 time: {:.10}s", PhillipsDessouky::profile_duration(profiling_start, profiling_end));
+        info!("PROFILING Rust_PhillipsDessouky::max_flow reformat OrderedFloat to f64 time: {:.10}s", utils::profile_duration(profiling_start, profiling_end));
 
         Ok(flows_f64)
-    }
-}
-
-// Private to Rust, not exposed to Python
-impl PhillipsDessouky {
-    fn profile_duration(start: Instant, end: Instant) -> f64 {
-        let duration: Duration = end.duration_since(start);
-        let seconds = duration.as_secs();
-        let subsec_nanos = duration.subsec_nanos();
-
-        let fractional_seconds = subsec_nanos as f64 / 1_000_000_000.0;
-        let total_seconds = seconds as f64 + fractional_seconds;
-
-        return total_seconds;
     }
 }
